Refactor module structure: rename 'solidspy' to 'src' and update import paths
