+++ docs/conf.py
@@ -21,7 +21,7 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('../solidspy'))
 
-import solidspy
+import src
 
 
 # -- General configuration ------------------------------------------------
