+++ src/solutil.py
